FinanceAPI/requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.6.1
future==0.17.1
idna==2.8
pycparser==2.19
pymongo==3.8.0
python-telegram-bot==11.1.0
PyYAML==5.1.1
requests>=2.7.0
six==1.12.0
telebot==0.0.3
urllib3==1.25.2
Django~=2.0.6
flask==1.1.1

